Revert "Update American-Pie.cs"
This reverts commit c21551657bbfb2c1ef9ca6d276eb00095528f6c4.

+++ American-Pie.cs
@@ -6,7 +6,7 @@ namespace FractionSums
     {
         static void Main()
         {
-            long firstNominator = long.Parse(Console.ReadLine());
+            int firstNominator = long.Parse(Console.ReadLine());
             long firstDenominator = long.Parse(Console.ReadLine());
 
             long secondNominator = long.Parse(Console.ReadLine());
